purchase and cart test added

+++ src/controllers/cart.controller.js
@@ -1,19 +1,39 @@
 const Cart = require("../models/Cart")
+const Category = require("../models/Category")
 const Product = require("../models/Product")
 const User = require("../models/User")
 const catchError = require("../utils/catchError")
 
 
 const getAllCarts = catchError(async (req, res) => {
-    const cart = await Cart.findAll({include: [Product, User]})
+    const userId = req.user.id
+
+    const cart = await Cart.findAll({
+        where: { userId },
+        include: [{
+            model: Product,
+            attributes: { exclude: ["createdAt", "updatedAt"] },
+            include: {
+                model: Category,
+                attributes: ["name"]
+            }
+        }]
+    })
 
     return res.json(cart)
 })
 
 const createCarts = catchError(async (req, res) => {
-    //const { name } = req.body;
+    const userId = req.user.id;
+    const { quantity, productId } = req.body;
 
-    const cartCreated = await Cart.create(req.body)
+    const newBody = {
+        quantity,
+        userId,
+        productId
+    }
+
+    const cartCreated = await Cart.create(newBody)
 
     return res.status(201).json(cartCreated)
 })
@@ -21,12 +41,13 @@ const createCarts = catchError(async (req, res) => {
 const updateCart = catchError(async (req, res) => {
     const { id } = req.params;
     const { quantity } = req.body;
+    const userId = req.user.id
 
     const cart = await Cart.findByPk(id)
-    if(!cart) return res.status(404).send("Cart not found")
+    if (!cart) return res.status(403).send("Cart not found")
 
-    const cartUpdated = await Cart.update({ quantity} , {
-        where: { id },
+    const cartUpdated = await Cart.update({ quantity }, {
+        where: { id, userId },
         returning: true
     })
 
@@ -35,9 +56,10 @@ const updateCart = catchError(async (req, res) => {
 
 const deleteCart = catchError(async (req, res) => {
     const { id } = req.params;
+    const userId = req.user.id
 
-    const cartToDelete = await Cart.destroy({ where: { id }})
-    if(!cartToDelete) return res.status(404).send("Cart not found")
+    const cartToDelete = await Cart.destroy({ where: { id, userId } })
+    if (!cartToDelete) return res.status(404).send("Cart not found")
 
     return res.sendStatus(204)
 })
